fix: [ci-fix] split track-event into server/client modules to fix build
trackEvent (server) imported next/headers transitively via the Supabase
server client. When client components imported trackEventClient from the
same file, the bundler pulled next/headers into the client bundle,
failing the Vercel build.

Moved trackEvent to track-event-server.ts so client components only
import track-event.ts which has no server dependencies.
